% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{centroid_distances}
\alias{centroid_distances}
\title{Compute the distances of each triad from a set of centroids.}
\usage{
centroid_distances(normalizedMeanExpression, factorList = NULL)
}
\description{
Compute the distances of each triad from a set of centroids.
}
