% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_cols.R
\name{checkColumns}
\alias{checkColumns}
\title{Check a column exists in a data.frame / tibble.}
\usage{
checkColumns(required, present, tbl_type)
}
\description{
This is just an internal function to make code neater.
}
