% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_runs.R
\name{compute_runs}
\alias{compute_runs}
\title{Compute runs of a variable in a triad data table.}
\usage{
compute_runs(data, var)
}
\arguments{
\item{data}{A data frame or tibble that contains triad data.}
}
\description{
Compute runs of a variable in a triad data table.
}
