% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances_from_centroids}
\alias{distances_from_centroids}
\title{Compute the distances of each triad from a series of centroids}
\usage{
distances_from_centroids(normalizedMeanExpression, lvl)
}
\description{
The centroids used are:
}
\details{
1. The centroid for all triads (column means of a matrix of normalized)
}
