% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad.expression.R
\docType{data}
\name{expression}
\alias{expression}
\title{Wheat expression experiment data for 210 samples}
\format{
A data frame with 51108 rows and 210 variables:
\describe{
  \item{gene}{name of the gene}
  \item{Sample_*}{A column of gene expression values for the sample corresponding to the column name}
  ...
}
}
\usage{
expression
}
\description{
A data table containing expression levels of 51108 genes for 210 samples,
derived from RNA-seq experiments.
}
\keyword{datasets}
