% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad.expression.R
\docType{data}
\name{expression_data}
\alias{expression_data}
\title{An example gene expression data-set.}
\format{
A data frame with 102442 rows and 233 variables.
}
\usage{
data(expression_data)
}
\description{
A dataset containning gene expression results of 102,442 genes
from 232 samples.
}
\details{
Each row represents a single gene.
The column 'gene', has the name of the gene and all other columns
are separate samples for which the levels of expression of each gene
were measured.

Note: The data comes from many varieties of wheat, but in this data-set,
the gene names are all Chinese Spring based, for all varieties.
So for example, values in the data-set for gene "TraesCS3A02G200800",
from a sample taken from the variety ARI, simply correspond to the gene
in ARI that is the homologue of the Chinese Spring gene "TraesCS3A02G200800".
}
\keyword{datasets}
