% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterExpressionForFactors.R
\name{filterExpressionForFactors}
\alias{filterExpressionForFactors}
\title{Filter a gene expression data set to specific combination of factor levels, based on the metadata describing the samples.}
\usage{
filterExpressionForFactors(expression, metadata, factorList)
}
\arguments{
\item{expression}{A data frame of expression values in which each column is a sample, and each row is a single gene.}

\item{metadata}{A data frame containing metadata for each sample.}

\item{factorLevels}{A List that represents a dictionary of factor and level pairs.}
}
\description{
Filter a gene expression data set to specific combination of factor levels, based on the metadata describing the samples.
}
