% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_means_by_factors.R
\name{gene_expression_means_by_factors}
\alias{gene_expression_means_by_factors}
\title{Compute mean expression for each gene for every combination of factor levels.}
\usage{
gene_expression_means_by_factors(expression, metadata, factorList)
}
\arguments{
\item{expression}{A data frame of expression values in which each column is a sample, and each row is a single gene.}

\item{metadata}{A data frame containing metadata for each sample.}

\item{factorList}{A list of column names that exist in `metadata` (as strings) to use as a factors.}

\item{homology}{A data frame of metadata about each triad. Include which genes make them up on the A, B, and D genomes.}
}
\description{
Compute mean expression for each gene for every combination of factor levels.
}
