% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_means_for_factor_levels.R
\name{gene_expression_means_for_factor_levels}
\alias{gene_expression_means_for_factor_levels}
\title{Compute mean expression for each gene, for a single combination of factor levels.}
\usage{
gene_expression_means_for_factor_levels(expression, metadata, factorLevels)
}
\arguments{
\item{expression}{A data frame of expression values in which each column is a sample, and each row is a single gene.}

\item{metadata}{A data frame containing metadata for each sample.}

\item{factorLevels}{A List that represents a dictionary of factor and level pairs.}
}
\description{
Compute mean expression for each gene, for a single combination of factor levels.
}
