% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad.expression.R
\docType{data}
\name{gene_locations}
\alias{gene_locations}
\title{An example gene annotation data-set}
\format{
A tibble with 318649 rows and 8 variables
}
\usage{
data(gene_locations)
}
\description{
Each row corresponds to a single gene's location in a single genome of a variety of wheat.
}
\details{
\itemize{
  \item chr. chromosome the gene exists in
  \item start. the base position in the genome, of the first base in the gene
  \item end. the base position in the genome, of the final base in the gene
  \item strand. which strand the gene exists on. Either '+' or '-'
  \item gene. the variety specific name of the gene
  \item group_id. the ID of the triad the gene belongs in
  \item variety. the variety of wheat
  \item subgenome. the subgenome of wheat the gene exists in ('A', 'B', or 'D')
}
}
\keyword{datasets}
