% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad_balance.R
\name{mk_expression_matrix}
\alias{mk_expression_matrix}
\title{Make a matrix from a table of normalised mean expression values}
\usage{
mk_expression_matrix(normalizedMeanExpression)
}
\description{
Putting the data into a matrix form makes it easy to do distances
from centroids etc.
}
