% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{normalize_triad_expression_means}
\alias{normalize_triad_expression_means}
\title{Normalize the triad mean expression data produced by triad_expression_mean_by_factors.}
\usage{
normalize_triad_expression_means(meanExpression, factorList = NULL)
}
\description{
Normalize the triad mean expression data produced by triad_expression_mean_by_factors.
}
