% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad_balance.R
\name{triad_balance}
\alias{triad_balance}
\title{Compute the expression balance from gene expression data.}
\usage{
triad_balance(normalizedMeanExpression, factorList = NULL)
}
\description{
Compute the expression balance from gene expression data.
}
