% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad.expression.R
\docType{data}
\name{wicker_te_regions}
\alias{wicker_te_regions}
\title{An example region dataset}
\format{
A data frame with 320 rows and 4 variables
}
\usage{
data(wicker_te_regions)
}
\description{
An example dataset of some genomic regions identified in this paper [todo].
}
\details{
The variables are:

\itemize{
  \item chr. chromosome the region exists in
  \item start. the base position in the genome of the first base of the region
  \item end. the base position in the genome of the final base of the region
  \item High.level.variety. The wheat variety this region was identified in
}
}
\keyword{datasets}
