% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{triad_expression_means_by_factors}
\alias{triad_expression_means_by_factors}
\title{Compute mean expression for each triad for each combination of factor levels.}
\usage{
triad_expression_means_by_factors(expression, metadata, homology, factorList)
}
\arguments{
\item{expression}{A data frame of expression values, in which each column is a sample, and each row is a single gene.}

\item{metadata}{A data frame containing metadata for each sample.}

\item{homology}{A data frame of metadata about each triad. Include which genes make them up on the A, B, and D genomes.}

\item{factorList}{A string of a metadata column name to use as a factor.}
}
\description{
Compute mean expression for each triad for each combination of factor levels.
}
